Progress So Far
The song_adder.py gets the user to log in and give access. They are then taken to the home page where they enter their playlist's URL. If it's valid, the songs in the 'list_of_songs' file procedurally get added to the target playlist (unless they're already in the playlist).